Code cleanup, some new "name" query parameters, OAuth in /metadata
* Removed some duplicate code from query parameter classes.
* Implemented a few new "name" query parameters.
* Added security OAuth entry to CapabilityStatement (/metadata) if OAuth
is enabled.
* Simplified some lambda expressions and switched to methods handles
where possible.
* Removed not needed casts.
* Removed explicit generic types, where types can be inferred.
* Combined some if/else-if statements.
* Added missing @Override annotations.
* Added final to class fields where applicable.
* Introduced switch-expressions where possible, remove default case in
switch-expressions if all cases covered.
* Regenerated equals and hashCode implementation based on Objects...
methods.
* Other minor code cleanup.

+++ dsf-fhir/dsf-fhir-server/src/test/java/dev/dsf/fhir/dao/AbstractResourceDaoTest.java
@@ -276,7 +276,7 @@ public abstract class AbstractResourceDaoTest<D extends Resource, C extends Reso
 		boolean deleted = dao.delete(UUID.fromString(createdResource.getIdElement().getIdPart()));
 		assertTrue(deleted);
 
-		D updatedResource = dao.update(updateResource(createdResource), (long) ResourceDao.FIRST_VERSION + 1L);
+		D updatedResource = dao.update(updateResource(createdResource), ResourceDao.FIRST_VERSION + 1L);
 		assertNotNull(updatedResource);
 		assertNotNull(updatedResource.getIdElement());
 		assertNotNull(updatedResource.getIdElement().getIdPart());
